+++ TestSutraUsers/src/main/java/com/tathvatech/survey/oid/SectionResponseOID.java
@@ -1,9 +1,10 @@
 package com.tathvatech.survey.oid;
 
-import com.tathvatech.ts.core.common.EntityTypeEnum;
-import com.tathvatech.ts.core.common.OID;
 
-public class SectionResponseOID extends OID{
+import com.tathvatech.common.enums.EntityTypeEnum;
+import com.tathvatech.user.OID.OID;
+
+public class SectionResponseOID extends OID {
 
 	public SectionResponseOID(int pk)
 	{
@@ -37,7 +38,7 @@ public class SectionResponseOID extends OID{
 	
 	@Override
 	public int hashCode() {
-		return super.getPk();
+		return (int) super.getPk();
 	}
 
 }
